MM Step Fix
